requirements.txt
click
requests
colorlog
booby-ng
python-dxf

# If SSL Connections fails, install these depenencies
# pyOpenSSL ndg-httpsclient pyasn1